pczt: Consolidate pczt_to_tx_data into Pczt::extract_tx_data
Moves pczt_to_tx_data from the tx_data module into a method on Pczt, and
eliminates the tx_data module entirely. All three consumer roles
(signer, io_finalizer, tx_extractor) now uniformly wrap a common
ExtractError error type instead of reproducing it in the variants
of each of those error types.
